backend/requirements.txt
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
PyMySQL
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
a
jinja2
python-pptx
gtts
moviepy
aiofiles
